pyproject.toml
[project]
name = "pd-hsr-assignment-4"
version = "0.0.0"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.11.0",
    "numpy>=2.2.6",
    "opencv-python>=4.12.0.88",
    "pillow>=12.0.0",
    "timm>=1.0.20",
    "torch>=2.9.0",
    "torchvision>=0.24.0",
    "tqdm>=4.67.1",
    "transformers>=4.57.1",
]
